requirements.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.2
decorator==5.1.1
Django==4.0.4
djangorestframework==3.13.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.7.0
inflection==0.5.1
iniconfig==1.1.1
ipython==8.3.0
jedi==0.18.1
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.1.1
wcwidth==0.2.5
